null feature menu and network feature menu are moved to their own separate files
